requirements.txt
chainlit==0.7.0
numpy==1.25.2
openai==0.27.8
python-dotenv==1.0.0
wandb==0.15.11
chromadb
langchain
tiktoken
pandas
scipy
scikit-learn
ipykernel
matplotlib
plotly
deeplake
google-search-results

